Fixed Undefined index: multi_categories on line 754,820,882 multi_copy on line 878 multi_delete on line 815 multi_move on line 751 multi_products on line 785,850,912,927,939 Undefined variable: cInfo in File: on line 953

+++ admin/includes/modules/categories_view.php
@@ -748,10 +748,10 @@
               // --------------------
               // multi_move confirm
               // --------------------
-              if (xtc_not_null($_POST['multi_move'])) {
+              if (isset($_POST['multi_move']) && xtc_not_null($_POST['multi_move'])) {
                 $heading[]  = array('text' => '<b>' . TEXT_INFO_HEADING_MOVE_ELEMENTS . '</b>');
                 $contents[] = array('text' => '<table width="100%" border="0">');
-                if (is_array($_POST['multi_categories'])) {
+                if (isset($_POST['multi_categories']) && is_array($_POST['multi_categories'])) {
                   foreach ($_POST['multi_categories'] AS $multi_category) {
                     $category_query = xtc_db_query("SELECT c.categories_id,
                                                            cd.categories_name,
@@ -782,7 +782,7 @@
                   $category_tree = xtc_get_category_tree();
                 }
 
-                if (is_array($_POST['multi_products'])) {
+                if (isset($_POST['multi_products']) && is_array($_POST['multi_products'])) {
                   if ($current_category_id != 0) {
                     foreach ($_POST['multi_products'] AS $multi_product) {
                       $contents[] = array('text' => '<tr><td style="border-bottom: 1px solid #af417e; margin-bottom: 10px;" class="infoBoxContent"><b>' . xtc_get_products_name($multi_product) . '</b></td></tr>');
@@ -812,12 +812,12 @@
               // --------------------
               // multi_delete confirm
               // --------------------
-              if (xtc_not_null($_POST['multi_delete'])) {
+              if (isset($_POST['multi_delete']) && xtc_not_null($_POST['multi_delete'])) {
                 $heading[]  = array('text' => '<b>' . TEXT_INFO_HEADING_DELETE_ELEMENTS . '</b>');
                 $contents[] = array('text' => '<a class="button" href="javascript:SwitchCheckDeleteConfirm()" onclick="this.blur()">' . BUTTON_REVERSE_SELECTION . '</a>');
                 $contents[] = array('text' => '<table width="100%" border="0">');
 
-                if (is_array($_POST['multi_categories'])) {
+                if (isset($_POST['multi_categories']) && is_array($_POST['multi_categories'])) {
                   foreach ($_POST['multi_categories'] AS $multi_category) {
                     $category_query = xtc_db_query("SELECT c.categories_id,
                                                            cd.categories_name,
@@ -847,7 +847,7 @@
                   }
                 }
 
-                if (is_array($_POST['multi_products'])) {
+                if (isset($_POST['multi_products']) && is_array($_POST['multi_products'])) {
                   foreach ($_POST['multi_products'] AS $multi_product) {
                     $contents[] = array('text' => '<tr><td style="border-bottom: 1px solid #af417e; margin-bottom: 10px;" class="infoBoxContent"><b>' . xtc_get_products_name($multi_product) . '</b></td></tr>');
                     $product_categories_string = '';
@@ -875,11 +875,11 @@
               // --------------------
               // multi_copy confirm
               // --------------------
-              if (xtc_not_null($_POST['multi_copy'])) {
+              if (isset($_POST['multi_copy']) && xtc_not_null($_POST['multi_copy'])) {
                 $heading[]  = array('text' => '<b>' . TEXT_INFO_HEADING_COPY_TO . '</b>');
                 $contents[] = array('text' => '<a class="button" href="javascript:SwitchCheckCopyConfirm()" onclick="this.blur()">' . BUTTON_REVERSE_SELECTION . '</a>');
                 $contents[] = array('text' => '<table width="100%" border="0">');
-                if (is_array($_POST['multi_categories'])) {
+                if (isset($_POST['multi_categories']) && is_array($_POST['multi_categories'])) {
                   foreach ($_POST['multi_categories'] AS $multi_category) {
                     $category_query = xtc_db_query("SELECT c.categories_id,
                                                            cd.categories_name,
@@ -909,7 +909,7 @@
                   }
                 }
 
-                if (is_array($_POST['multi_products'])) {
+                if (isset($_POST['multi_products']) && is_array($_POST['multi_products'])) {
                   foreach ($_POST['multi_products'] AS $multi_product) {
                     $contents[] = array('text' => '<tr><td style="border-bottom: 1px solid #af417e; margin-bottom: 10px;" class="infoBoxContent"><b>' . xtc_get_products_name($multi_product) . '</b></td></tr>');
                     $product_categories_string = '';
@@ -924,7 +924,7 @@
                 if (QUICKLINK_ACTIVATED=='true') {
                   $contents[] = array('text' => '<hr noshade>');
                   $contents[] = array('text' => '<b>'.TEXT_MULTICOPY.'</b><br />'.TEXT_MULTICOPY_DESC);
-                  if (is_array($_POST['multi_products'])) {
+                  if (isset($_POST['multi_products']) && is_array($_POST['multi_products'])) {
                     $cat_tree=xtc_get_category_tree('0','','0');
                   } else {
                     $cat_tree=xtc_get_category_tree();
@@ -936,7 +936,7 @@
                   $contents[] = array('text' => $tree.'<br /><hr noshade>');
                   $contents[] = array('text' => '<b>'.TEXT_SINGLECOPY.'</b><br />'.TEXT_SINGLECOPY_DESC);
                 }
-                if (is_array($_POST['multi_products'])) {
+                if (isset($_POST['multi_products']) && is_array($_POST['multi_products'])) {
                   $category_tree=xtc_get_category_tree('0','','0');
                 } else {
                   $category_tree=xtc_get_category_tree();
@@ -949,8 +949,8 @@
                 $contents[] = array('text' => '<strong>' . TEXT_CONTENT_COPY . '</strong><br />' . xtc_draw_checkbox_field('cnt_copy', 'cnt_copy', false).'<font size="1">'.TEXT_CONTENT_COPY_INFO.'</font><br /><hr noshade>');
                 $contents[] = array('text' => '<strong>' . TEXT_LINKS_COPY . '</strong><br />' . xtc_draw_checkbox_field('links_copy', 'links_copy', false).'<font size="1">'.TEXT_LINKS_COPY_INFO.'</font><br /><hr noshade>');
                 // EOF - Timo Paul (mail[at]timopaul[dot]biz) - 2014-01-17 - duplicate products content and links
-                $contents[] = array('text' => '<strong>' . TEXT_TAGS_COPY . '</strong><br />' . xtc_draw_checkbox_field('tags_copy', 'tags_copy', false).'<font size="1">'.TEXT_TAGS_COPY_INFO.'</font><br /><hr noshade>');
-                $contents[] = array('align' => 'center', 'text' => '<input class="button" type="submit" name="multi_copy_confirm" value="' . BUTTON_COPY . '"> <a class="button" href="' . xtc_href_link(FILENAME_CATEGORIES, xtc_get_all_get_params(array('cPath', 'action', 'pID', 'cID')) . 'cPath=' . $cPath . '&pID=' . $pInfo->products_id . '&cID=' . $cInfo->categories_id) . '">' . BUTTON_CANCEL . '</a>');
+                $contents[] = array('text' => '<strong>' . TEXT_TAGS_COPY . '</strong><br />' . xtc_draw_checkbox_field('tags_copy', 'tags_copy', false).'<font size="1">'.TEXT_TAGS_COPY_INFO.'</font><br /><hr noshade>');                
+                $contents[] = array('align' => 'center', 'text' => '<input class="button" type="submit" name="multi_copy_confirm" value="' . BUTTON_COPY . '"> <a class="button" href="' . xtc_href_link(FILENAME_CATEGORIES, xtc_get_all_get_params(array('cPath', 'action', 'pID', 'cID')) . 'cPath=' . $cPath . '&pID=' . $pInfo->products_id . '&cID=' . (isset($cInfo->categories_id) ? $cInfo->categories_id : $current_category_id)) . '">' . BUTTON_CANCEL . '</a>');
               }
               // multi_copy confirm ENDS
               break;
